Refactor 2023 day 3

+++ 2023/day3/main_test.go
@@ -39,7 +39,7 @@ func TestSchematic_PartTwo(t *testing.T) {
 	tests := []struct {
 		name   string
 		input  string
-		expect int64
+		expect int
 	}{
 		{
 			name: "Example 1",
